countChildrenを別ファイルに分離
